# Optimize method of hasClass