这里我使用了 jQuery hasClass 的技巧。
能够大幅提升 hasClass 的判断速度。

希望能够采纳这个PR。

这里我没有检测obj.className不过加上之后 速度几乎没有影响
原本的方式开销主要在于 split 方法。

```JavaScript
return obj.className ? ((" " + obj.className + " ").indexOf(" " + classStr + " ") > -1) : false;
```

```JavaScript
var obj = {
    className: "a b c aa bb cc dd test ee a bb cc"
}
console.time("Master");
for(let i = 0; i < 1000000; i++){
     ecDo.hasClass(obj, "test");
}
console.timeEnd("Master"); // 950ms ~ 1050ms

console.time("Hongxu");
for(let i = 0; i < 1000000; i++){
     ecDo.myHasClass(obj, "test");
}
console.timeEnd("Hongxu"); // 25ms ~ 30ms
```