@chenhuiYj 2018-04-24
我测试了下，不判断obj.className也没事，至少这个也是一个空白字符串。能正常运行。感谢你的建议，现在合并了