@lynxerzhang 2018-05-16
想请教下，obj.className的前后都添加了空格，是不是为了兼容性？不然不加空格也是可以的