@hongxuWei 2018-05-17
是因为这里 `indexOf(" " + classStr + " ")` 前后加了空格。

如果 className 的前后不加空格，那么 className 的 第一个和最后一个就不会匹配到了。

```JavaScript
var obj = {
    className: "aa1 bb2 cc3"
}
 ecDo.hasClass(obj, "aa1"); // false

// 如果className 前后不加空格相当于

"aa1 bb2 cc3".indexOf( " aa1 " );   // -1
```

具体可以看下 jQuery hasClass 部分的源码。

我的 pr 也比较简单，没有对 输入的 className 进行 strip 以及多个参数的支持。