@lynxerzhang 2018-05-21
哦..对，明白了，我现在在查看JQuery hasClass的源码，感谢回复！